requirements.txt
# scraping
selenium==4.43.0
webdriver-manager==4.0.2
requests==2.33.1
beautifulsoup4==4.14.3

# data processing
pandas==3.0.2
numpy==2.4.4

# database
sqlalchemy==2.0.49
psycopg2-binary==2.9.12

# env
python-dotenv==1.2.2

# geocoding
geopy==2.4.1